Refactor les events dans un module
