+++ src/modules/events/constants.ts
